+++ quality_check/config.py
@@ -42,6 +42,7 @@ KEY_MAX_TAGS = 'maxTags'
 KEY_MAX_TAG_EXCLUSIONS = 'maxTagExclusions'
 KEY_HIDDEN_MENUS = 'hiddenMenus'
 KEY_SEARCH_SCOPE = 'searchScope'
+KEY_SEARCH_HISTORY_LENGTH = 'searchHistoryLength'
 
 SCOPE_LIBRARY = 'Library'
 SCOPE_SELECTION = 'Selection'
@@ -50,6 +51,7 @@ DEFAULT_STORE_VALUES = {
                            KEY_MAX_TAGS: 5,
                            KEY_MAX_TAG_EXCLUSIONS: [],
                            KEY_HIDDEN_MENUS: [],
+                           KEY_SEARCH_HISTORY_LENGTH: 10
                        }
 
 # Per library we store an exclusions map
@@ -328,6 +330,15 @@ class ConfigWidget(QWidget):
         if c.get(KEY_SUPPRESS_FIX_DIALOG, False):
             self.suppress_dialog_checkbox.setCheckState(Qt.Checked)
         other_layout.addWidget(self.suppress_dialog_checkbox, 1, 0, 1, 2)
+     
+        search_history_label = QLabel(_('Search history count:'), self)
+        search_history_label.setToolTip(_('How many previous search expressions to remember on Search sPubs dialog'))
+        other_layout.addWidget(search_history_label, 2, 0, 1, 1)
+        self.search_history_spin = QSpinBox(self)
+        self.search_history_spin.setMinimum(0)
+        self.search_history_spin.setMaximum(100)
+        self.search_history_spin.setProperty('value', c.get(KEY_SEARCH_HISTORY_LENGTH, 10))
+        other_layout.addWidget(self.search_history_spin, 2, 1, 1, 1)
         other_layout.setColumnStretch(2, 1)
 
         menus_groupbox = QGroupBox(_('Visible menus'))
@@ -363,7 +374,8 @@ class ConfigWidget(QWidget):
         new_prefs[KEY_AUTHOR_INITIALS_MODE] = self.initials_combo.selected_key()
         new_prefs[KEY_SUPPRESS_FIX_DIALOG] = self.suppress_dialog_checkbox.checkState() == Qt.Checked
         new_prefs[KEY_SEARCH_SCOPE] = plugin_prefs[STORE_OPTIONS].get(KEY_SEARCH_SCOPE, SCOPE_LIBRARY)
-
+        new_prefs[KEY_SEARCH_HISTORY_LENGTH] = int(unicode(self.search_history_spin.value()))
+        
         new_prefs[KEY_HIDDEN_MENUS] = self.visible_menus_list.get_hidden_menus()
         # For each menu that was visible but now is not, we need to unregister any
         # keyboard shortcut associated with that action.